feat: extract central bodies data to shared header
